requirements.txt
3==0.0.0
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
cssselect2==0.6.0
distlib==0.3.5
Django==4.1
django-ckeditor==6.5.0
django-cors-headers==3.8.0
django-environ==0.4.5
django-js-asset==2.0.0
django-mssql-backend==2.8.1
django-pyodbc-azure==2.1.0.0
django-rest-knox==4.2.0
djangorestframework==3.12.4
drf-extra-fields==3.4.0
drf-writable-nested==0.7.0
filelock==3.8.0
fonttools==4.38.0
future==0.18.2
html5lib==1.1
idna==3.4
lxml==4.9.1
mysqlclient==2.1.1
oscrypto==1.3.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
pydyf==0.5.0
pyHanko==0.14.0
pyhanko-certvalidator==0.19.5
pyodbc==4.0.34
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tinycss2==1.1.1
tqdm==4.64.1
tzdata==2022.2
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
virtualenv==20.16.3
webencodings==0.5.1
xhtml2pdf==0.2.8
zopfli==0.2.1